requirements.txt
pandas
numpy
pyarrow
openpyxl
duckdb
tabulate
pypdf
llama-cpp-python

